+++ WellDesignedAPI.Common/Models/Request/FilterState.cs
@@ -3,9 +3,8 @@
     public class FilterState
     {
         public string PropertyName { get; set; }
-        public string Value { get; set; }
-        public string? RangeFrom { get; set; }
-        public string? RangeTo { get; set; }
+        public string? ValueFromOrEqualTo { get; set; }
+        public string? ValueTo { get; set; }
     }
 
 }